refactor(frontend): sharpen agent studio shell submodels
