+++ packages/frontend/src/pages/agents/shell/use-agent-studio-selection-intent-state.ts
@@ -1,0 +1,82 @@
+import type { AgentRole } from "@openducktor/core";
+import { useCallback, useEffect, useState } from "react";
+import {
+  type AgentStudioSelectionIntent,
+  isSelectionIntentResolved,
+} from "./agent-studio-selection-intent";
+
+type UseAgentStudioSelectionIntentStateArgs = {
+  isRepoNavigationBoundaryPending: boolean;
+  taskIdParam: string;
+  sessionParam: string | null;
+  roleFromQuery: AgentRole;
+};
+
+export type AgentStudioSelectionIntentState = {
+  selectionIntentForController: AgentStudioSelectionIntent | null;
+  isSessionSelectionResolving: boolean;
+  scheduleSelectionIntent: (intent: AgentStudioSelectionIntent) => void;
+};
+
+export function useAgentStudioSelectionIntentState({
+  isRepoNavigationBoundaryPending,
+  taskIdParam,
+  sessionParam,
+  roleFromQuery,
+}: UseAgentStudioSelectionIntentStateArgs): AgentStudioSelectionIntentState {
+  const [selectionIntent, setSelectionIntent] = useState<AgentStudioSelectionIntent | null>(null);
+  const [sessionlessSelection, setSessionlessSelection] =
+    useState<AgentStudioSelectionIntent | null>(null);
+
+  const scheduleSelectionIntent = useCallback((intent: AgentStudioSelectionIntent): void => {
+    setSelectionIntent(intent);
+    setSessionlessSelection(intent.externalSessionId === null ? intent : null);
+  }, []);
+
+  const activeSessionlessSelection =
+    sessionlessSelection &&
+    sessionlessSelection.taskId === taskIdParam &&
+    sessionlessSelection.role === roleFromQuery &&
+    sessionParam === null
+      ? sessionlessSelection
+      : null;
+
+  useEffect(() => {
+    if (isRepoNavigationBoundaryPending) {
+      setSelectionIntent(null);
+      return;
+    }
+
+    if (!selectionIntent) {
+      return;
+    }
+
+    const selectionIntentResolved = isSelectionIntentResolved({
+      selectionIntent,
+      taskIdParam,
+      sessionParam,
+      roleFromQuery,
+    });
+
+    if (selectionIntentResolved) {
+      setSelectionIntent(null);
+    }
+  }, [isRepoNavigationBoundaryPending, roleFromQuery, selectionIntent, sessionParam, taskIdParam]);
+
+  const isSessionSelectionResolving = Boolean(
+    selectionIntent &&
+      !isRepoNavigationBoundaryPending &&
+      !isSelectionIntentResolved({
+        selectionIntent,
+        taskIdParam,
+        sessionParam,
+        roleFromQuery,
+      }),
+  );
+
+  return {
+    selectionIntentForController: selectionIntent ?? activeSessionlessSelection,
+    isSessionSelectionResolving,
+    scheduleSelectionIntent,
+  };
+}
