+++ packages/frontend/src/pages/agents/shell/use-agents-page-route-session-model.ts
@@ -1,6 +1,6 @@
 import type { RuntimeDescriptor, RuntimeKind } from "@openducktor/contracts";
 import type { AgentModelCatalog, AgentRole, AgentSessionTodoItem } from "@openducktor/core";
-import { useCallback, useEffect, useRef, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { useNavigationType, useSearchParams } from "react-router-dom";
 import type { AgentSessionSummary } from "@/state/agent-sessions-store";
 import type { AgentSessionState } from "@/types/agent-orchestrator";
@@ -10,10 +10,8 @@ import { useAgentStudioQuerySessionSync } from "../use-agent-studio-query-sessio
 import { useAgentStudioQuerySync } from "../use-agent-studio-query-sync";
 import { useAgentStudioSelectionController } from "../use-agent-studio-selection-controller";
 import { useAgentStudioReadiness } from "../use-agents-page-readiness";
-import {
-  type AgentStudioSelectionIntent,
-  isSelectionIntentResolved,
-} from "./agent-studio-selection-intent";
+import { useAgentStudioSelectionIntentState } from "./use-agent-studio-selection-intent-state";
+import { useAgentStudioWorktreeRecoverySignal } from "./use-agent-studio-worktree-recovery-signal";
 
 type UseAgentsPageRouteSessionModelArgs = {
   activeWorkspace: ActiveWorkspace | null;
@@ -92,11 +90,6 @@ export function useAgentsPageRouteSessionModel({
   const [searchParams, setSearchParams] = useSearchParams();
   const navigationType = useNavigationType();
   const [contextSwitchVersion, setContextSwitchVersion] = useState(0);
-  const [selectionIntent, setSelectionIntent] = useState<AgentStudioSelectionIntent | null>(null);
-  const [sessionlessSelection, setSessionlessSelection] =
-    useState<AgentStudioSelectionIntent | null>(null);
-  const [worktreeRecoverySignal, setWorktreeRecoverySignal] = useState(0);
-  const lastWorktreeRecoveryKeyRef = useRef<string | null>(null);
 
   const {
     taskIdParam,
@@ -125,21 +118,13 @@ export function useAgentsPageRouteSessionModel({
     setContextSwitchVersion((current) => current + 1);
   }, []);
 
-  const scheduleSelectionIntent = useCallback(
-    (intent: { taskId: string; externalSessionId: string | null; role: AgentRole }): void => {
-      setSelectionIntent(intent);
-      setSessionlessSelection(intent.externalSessionId === null ? intent : null);
-    },
-    [],
-  );
-
-  const activeSessionlessSelection =
-    sessionlessSelection &&
-    sessionlessSelection.taskId === taskIdParam &&
-    sessionlessSelection.role === roleFromQuery &&
-    sessionParam === null
-      ? sessionlessSelection
-      : null;
+  const { selectionIntentForController, isSessionSelectionResolving, scheduleSelectionIntent } =
+    useAgentStudioSelectionIntentState({
+      isRepoNavigationBoundaryPending,
+      taskIdParam,
+      sessionParam,
+      roleFromQuery,
+    });
 
   const readiness = useAgentStudioReadiness({
     activeWorkspace,
@@ -161,7 +146,7 @@ export function useAgentsPageRouteSessionModel({
     sessionParam,
     hasExplicitRoleParam,
     roleFromQuery,
-    selectionIntent: selectionIntent ?? activeSessionlessSelection,
+    selectionIntent: selectionIntentForController,
     updateQuery: scheduleQueryUpdate,
     agentStudioReadinessState: readiness.agentStudioReadinessState,
     hydrateRequestedTaskSessionHistory,
@@ -173,74 +158,11 @@ export function useAgentsPageRouteSessionModel({
     onContextSwitchIntent: signalContextSwitchIntent,
   });
 
-  useEffect(() => {
-    if (isRepoNavigationBoundaryPending) {
-      setSelectionIntent(null);
-      return;
-    }
-
-    if (!selectionIntent) {
-      return;
-    }
-
-    const selectionIntentResolved = isSelectionIntentResolved({
-      selectionIntent,
-      taskIdParam,
-      sessionParam,
-      roleFromQuery,
-    });
-
-    if (selectionIntentResolved) {
-      setSelectionIntent(null);
-    }
-  }, [isRepoNavigationBoundaryPending, roleFromQuery, selectionIntent, sessionParam, taskIdParam]);
-
-  const isSessionSelectionResolving = Boolean(
-    selectionIntent &&
-      !isRepoNavigationBoundaryPending &&
-      !isSelectionIntentResolved({
-        selectionIntent,
-        taskIdParam,
-        sessionParam,
-        roleFromQuery,
-      }),
-  );
-
-  useEffect(() => {
-    const nextRecoveryKey = [
-      workspaceRepoPath ?? "",
-      selection.viewTaskId ?? "",
-      selection.viewSelectedTask?.updatedAt ?? "",
-      selection.viewSelectedTask?.status ?? "",
-      selection.viewActiveSession?.externalSessionId ?? "",
-      selection.viewActiveSession?.status ?? "",
-      selection.viewActiveSession?.workingDirectory ?? "",
-      selection.isViewSessionHistoryHydrating ? "1" : "0",
-      isForegroundLoadingTasks ? "1" : "0",
-    ].join(":");
-
-    if (lastWorktreeRecoveryKeyRef.current === null) {
-      lastWorktreeRecoveryKeyRef.current = nextRecoveryKey;
-      return;
-    }
-
-    if (lastWorktreeRecoveryKeyRef.current === nextRecoveryKey) {
-      return;
-    }
-
-    lastWorktreeRecoveryKeyRef.current = nextRecoveryKey;
-    setWorktreeRecoverySignal((previous) => previous + 1);
-  }, [
-    isForegroundLoadingTasks,
-    selection.isViewSessionHistoryHydrating,
-    selection.viewActiveSession?.externalSessionId,
-    selection.viewActiveSession?.status,
-    selection.viewActiveSession?.workingDirectory,
-    selection.viewSelectedTask?.status,
-    selection.viewSelectedTask?.updatedAt,
-    selection.viewTaskId,
+  const worktreeRecoverySignal = useAgentStudioWorktreeRecoverySignal({
     workspaceRepoPath,
-  ]);
+    selection,
+    isForegroundLoadingTasks,
+  });
 
   useEffect(() => {
     if (!workspaceRepoPath || runtimeDefinitions.length === 0 || isLoadingChecks) {
